% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_plot.R
\name{categoric_plot}
\alias{categoric_plot}
\title{Categorical Variables Plots}
\usage{
categoric_plot(
  data,
  pie = TRUE,
  pie_legend = TRUE,
  pie_legend_size = 0.5,
  pie_legend_position = "bottom",
  pie_inset = c(0, -0.15),
  bar = TRUE,
  bar_legend = TRUE,
  bar_legend_size = 0.5,
  bar_legend_position = "bottom",
  bar_inset = c(0, -0.4),
  n_col = 1,
  bar_width = 0.8,
  bar_height = NULL
)
}
\arguments{
\item{data}{input data}

\item{pie}{a logical argument(default TRUE) that determines if pie plot is generated}

\item{pie_legend}{a logical argument(default TRUE) that determines if a legend of pie is generated}

\item{pie_legend_size}{an argument(default is 0.5) that determines size of the legend}

\item{pie_legend_position}{an argument(dafult is bottom) that determines the position of the legend}

\item{bar}{a logical argument(default TRUE) that determines if bar plot is generated}

\item{bar_legend}{a logical argument(default TRUE) that determines if a legend of bar is generated}

\item{bar_legend_size}{an argument(default is 0.5) that determines size of the legend}

\item{bar_legend_position}{an argument(dafult is bottom) that determines the position of the legend}

\item{n_col}{an argument that determine how many plot being put on the same rows}

\item{bar_width}{width of bar plot}

\item{bar_height}{height of bar plot}
}
\value{
a list of pie charts
}
\description{
Categorical Variables Plots
}
\examples{
categoric_plot(data(mtcars))
}
