% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_cluster.R
\name{corr_cluster}
\alias{corr_cluster}
\title{Correlation Cluster}
\usage{
corr_cluster(data, label = NULL, type = "pearson", method = "complete")
}
\arguments{
\item{data}{input dataset}

\item{label}{label or name for the dataset, optional.}

\item{type}{pearson or spearman correlation, default is pearson}

\item{method}{the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC). Dafault is complete}
}
\value{
A plot of dendrogram
}
\description{
corr_cluster performs hierarchical clustering based on the corrleation distance, and generates a dendrogram.
}
\examples{
corr_cluster(data(mtcars),type='pearson',method='complete')
}
