% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_impute.R
\name{impute_missing}
\alias{impute_missing}
\title{Missing Value Imputation}
\usage{
impute_missing(data, method = "mean", k = NULL)
}
\arguments{
\item{data}{input data}

\item{method}{method of handling missing values: mean, median, mode, locf,knn.}

\item{k}{argument used in knn method. k is value of number of neighbors will be checked. Default is Null}
}
\description{
Missing Value Imputation
}
\examples{
impute_missing(data(mtcars), method='mean')
}
