% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_plot.R
\name{numeric_plot}
\alias{numeric_plot}
\title{Numerical Variables Distribution}
\usage{
numeric_plot(data, hist = TRUE, prob = FALSE, dens = FALSE)
}
\arguments{
\item{data}{input data}

\item{hist}{a logical argument(default TRUE) that determines if histogram is generated}

\item{prob}{a logical argument(default FALSE) that determines it is a probability histogram or relative frequency histogram.}

\item{dens}{a logical argument(default FALSE) that determine if density line is generated}
}
\value{
a list of plots
}
\description{
Numerical Variables Distribution
}
