% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_plot.R
\name{pie_plot}
\alias{pie_plot}
\title{Pie Plots for Categorical Variables}
\usage{
pie_plot(data, html = FALSE)
}
\arguments{
\item{data}{The input data frame containing categorical variables.}

\item{html}{Whether the output should be in HTML format,used when knitting into HTML. Default is FALSE.}
}
\value{
A list of pie charts.
}
\description{
This function generates pie charts for categorical variables in the input data frame using plotly.
Pie plots offer a visual representation of the distribution of categorical variables,
making it easy to understand the frequency of each category. They are particularly
useful for exploring patterns, identifying dominant categories, and comparing the relative
frequencies of different levels within each variable.
}
\examples{
data(iris)
pie_plot(iris)
}
