% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr_matrix}
\alias{corr_matrix}
\title{Correlation Matrix}
\usage{
corr_matrix(
  data,
  type = "pearson",
  decreasing = TRUE,
  corrplot = FALSE,
  order = "hclust",
  sig.level = 0.01,
  addCoef.col = NULL,
  number.cex = NULL
)
}
\arguments{
\item{data}{input dataset}

\item{type}{pearson or spearman correlation, default is pearson}

\item{decreasing}{sort the data frame by absolute correlation values, default is sorted by decreasing value}

\item{corrplot}{generate a correlation matrix plot, default is true}

\item{order}{Character, the ordering method of the correlation matrix.
'hclust' for the hierarchical clustering order (default).         
'original' for the original order.       
'AOE' for the angular order of the eigenvectors.   
'FPC' for the first principal component order.    
'alphabet' for alphabetical order.}

\item{sig.level}{Significant level. default is 0.01. If correlation p-value is less than 0.01, wipe away the corresponding glyph.}

\item{addCoef.col}{Color of coefficients added on the graph, if NULL(dafult), add no coefficients}

\item{number.cex}{Size of the coefficents values}
}
\value{
a data frame which contain rows names, column names, correlation coefficients, and p-values

a plot of correlation if corrplot is set to be true
}
\description{
Column 1: row names (variable 1 for the correlation test).   
  
  Column 2: column names (variable 2 for the correlation test).    
  
  Column 3: the correlation coefficients.    
  
  Column 4: the p-values of the correlations.
}
\examples{
corr_matrix(data(mtcars),type='pearson')
}
